fix hover-card
